+++ dio/CHANGELOG.md
@@ -8,6 +8,7 @@ See the [Migration Guide][] for the complete breaking changes list.**
 - Raise warning for `Map`s other than `Map<String, dynamic>` when encoding request data.
 - Improve exception messages
 - Allow `ResponseDecoder` and `RequestEncoder` to be async
+- Fix cancellation and missing progress handling for streamed responses and downloads when using `IOHttpClientAdapter`.
 
 ## 5.3.3